chore: put temp csv import script stuff into separate package
