+++ cmd/app/cli/script/tempcsvimport/battle_csv_import.go
@@ -1,4 +1,4 @@
-package script
+package tempcsvimport
 
 import (
 	"log"
