+++ cmd/app/cli/script/tempcsvimport/column_handler.go
@@ -1,4 +1,4 @@
-package script
+package tempcsvimport
 
 import (
 	"strconv"
